Update FPS_Blink.ino
- added a description for this example
- updated the hardware hookup
- added option to adjust software serial for other Arduino microcontrollers (i.e. ATmega32U4, ATmega2560, etc.) due to limitations with the library.
- added some a little bit more comments. =)
- changed resistors used in voltage division:

__Logic Level Conversion__

I tested the setup with the previous resistors. I was not able to get it working with the GT511C1R model. However, I was able to get it working when using a voltage divider (with 10kΩ and 20kΩ) similar to the setup used in our previous logic level converter’s application circuit [ [https://cdn.sparkfun.com/assets/b/0/e/1/0/522637c6757b7f2b228b4568.png](https://cdn.sparkfun.com/assets/b/0/e/1/0/522637c6757b7f2b228b4568.png) ]. I believe the IC used in the GT511C1R requires a slightly higher logic level for the input voltage in order for the signal to be considered a logic HIGH by the fingerprint scanner [similar to this general explanation of 3.3V CMOS Logic Levels – [https://learn.sparkfun.com/tutorials/logic-levels#33-v-cmos-logic-levels(https://learn.sparkfun.com/tutorials/logic-levels#33-v-cmos-logic-levels) ].

Doing a stress test (with the original 560Ohm and 1000Ohm resistor) and measuring the voltage with a multimeter, I noticed that the divided voltage was about 3.14V. When the GT511C1R was connected, the voltage dropped to about 2.97V. This is probably not high enough for the fingerprint scanner to recognize a logic HIGH.

_Replace Resistors for Higher Logic Level Threshold_

I was able to get the Arduino example code working by replacing the 560Ohm resistor with the 10kΩ resistor and replacing the 1000Ohm resistor with two 10kΩ resistors in series for 20kΩ. Testing the circuit with the resistors replaced, my multimeter read a divided voltage of 3.3V. When the GT511C1R was connected, the voltage dropped to about ~3.18V-3.04V. This was sufficient enough for the fingerprint scanner to receive the commands from the Arduino. I tested this with the blink examples and there were no problems.